Add a test for debouncing.
The test works by having a subscription to the output topic.
Every time the subscription is received, we increment a counter.
There is also a timer callback that executes every 0.1 seconds,
publishing the 'joy' message.  Like other tests, this publication
toggles the button on and off to avoid debouncing.  Unlike other
tests, once we have seen one message come in, we stop toggling
and just set it to one all of the time to ensure debouncing
works properly.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>