Add the ability to have deadman axes. (#46)
* Add the ability to have deadman axes.

Some controllers don't have a convenient shoulder trigger
button, but do have shoulder "axes".  Allow the axes to
be used for a deadman trigger, assuming they are pressed
all the way.  Note that I used a dict for the list of
axes, as this provides the most convenient way to deal
with controllers that use 1.0, -1.0, or 0.0 as the "far"
end of the axis.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Make sure to ignore buttons and axes that don't exist.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>